Refactor JupyterLab deployment and add support for oct_upgrade image
This commit refactors the `deploy_notebook` function in `jupyterlab.py` to include support for the `oct_upgrade` image. It adds a new condition to set the `start_script` variable to `/ML_platform_tests/SetupJupyterLab.sh` when the `image` contains `oct_upgrade`. Additionally, the `supported_images` function in the same file is updated to include the `oct_upgrade` image.

In the `pod.yaml` file, the `SetupPrivateJupyterLab.sh` script is replaced with the `start_script` variable, allowing for dynamic script selection based on the image.

Lastly, the `jupyterlab_form.html` template is modified to include the `ml-platform:oct_upgrade` option in the image dropdown menu.

These changes enhance the flexibility of JupyterLab deployment and provide support for the `oct_upgrade` image.